Add TypeIntrospection module with helper for getting names for types
A handful of the expansion templates I've written so far need a way to
handle optional fields, collections, etc., and need to know how to write
their type name in the resulting code. This new module contains an
improved version of the code I've used thus far, and provides short
names for things like `int` and `bool` that have built-in aliases in F#,
as well as nice handling of e.g. F# functions like `int -> int` vs
`FSharpFunc<int,int>`.