Add a paket reference to FSharp.Core
First thing of note is that I cannot seem to find a copy of FSharp.Core
with the .optdata and .sigdata files on the installation of Visual
Studio for Mac. While I do not remember why these were needed, this is
the fastest way to bypass their non-existence (or their non-obvious
location).