Add FSharp.Core detection when installed by Paket or NuGet
This is roughly based on the assumption that
`typeof<CompileSource>.Assembly.Location` will point to the copy of this
type provider from inside the `packages` directory that Paket/NuGet
creates, and that FSharp.Core will be there as well.

In the future, for NuGet path handling, may want to create a helper
function that creates different paths for different versions? Not a big
deal when installed by Paket, since the path used to find FSharp.Core
will not contain the package version number.