Pass the new `--nocopyfsharpcore` parameter to fsc.exe
When running tests in the templates project, F# Interactive and fsc both
attempt to put a copy of `FSharp.Core.dll` in `%TEMP%`, which fails for
fsc because the file is already locked by F# Interactive. Passing this
flag to fsc lets it know not to copy our version of `FSharp.Core`
anywhere.