Update paket.bootstrapper.exe to v5.142.0 to handle GitHub's TLS changes
Matching @forki's automated GitHub pull request on the templates
repository, this updated version is required due to GitHub changing its
servers to disallow connections that do not use TLS v1.2+. Without this
update, the bootstrapper is unable to download `paket.exe` during the
build process.

This commit uses the same copy of `paket.bootstrapper.exe` as the
templates project.