Omit temporary source file when generating an assembly with no expansion
When the `CreateAssembly` output option is in use, and the active type
expansion templates (if any are actually active) do not produce output,
the type provider gets a build error from the F# compiler indicating
that the temporary source file does not contain anything useful.

This commit causes the temporary source file to not be generated, and no
longer asks the F# compiler to use it when generating the resulting
assembly. Although producing an output that does not contain any
expanded code somewhat defeats the purpose of using this type provider,
it is possible that someone unfamiliar with the type provider will
remove or comment out some code and end up with a build error they don't
understand; this commit improves that developer's day just a little bit.

Closes #20