Add possible F# compiler paths, including those for Visual Studio 2022
While testing a work project, found that the current version of Visual
Studio 2019 doesn't drop `fsc.exe` in the same path that it used to. One
of the other fallback paths has probably been working for me.

Going forward, the locations used for Visual Studio 2022 are at the top
of the list, so they'll be used first. This won't be useful for
everyone, but it'll open up the ability to use newer language features
in the code being processed.