Reference user project version of FSharp.Core in the temp dummy assembly
While testing the last release of this project against Visual Studio
2019, found that I was getting errors that FSharp.Core version 6 could
not be found while loading the dummy assembly. This didn't affect
building in Visual Studio 2022, which likely uses version 6 internally
for a handful of things.

Changing the dummy assembly to reference the user project's version of
FSharp.Core seems to work fine in both 2019 and 2022, so proceeding with
that.